+++ taxel-gui/CHANGELOG.md
@@ -6,6 +6,7 @@
   - Fix non-nil facts with empty string as value
   - Fix unsanitized xml for `Eric::validate`
   - Fix missing `context_ref` on tuple child after editing via dropdown
+  - Allow search on enabled report sections only
 - added
   - Import report
   - Display report